feat: add support for Kotlinx Kover and enhance OTP handling
- Integrate Kotlinx Kover for code coverage
- Add C interop for zlib
- Implement new functions for HOTP and TOTP generation from URIs
- Update TOTP and HOTP classes to use Long for time intervals and periods
- Refactor GitHub Actions to clarify test steps for different platforms

+++ sharedLib/src/commonMain/kotlin/tech/arnav/twofac/lib/otp/TOTP.kt
@@ -19,7 +19,7 @@ class TOTP(
      * Time interval in seconds for TOTP generation.
      * Default is 30 seconds, which is the standard for TOTP.
      */
-    val timeInterval: Int = 30
+    val timeInterval: Long = 30
 ) : OTP {
 
     // Use HOTP internally for the actual OTP generation
